+++ golang/lib/testsuite/test_configuration_builder.go
@@ -32,7 +32,6 @@ type TestConfigurationBuilder struct {
 	setupTimeoutSeconds uint32
 	runTimeoutSeconds uint32
 	isPartioningEnabled bool
-	staticFileFilepaths map[services.StaticFileID]string
 	filesArtifactUrls map[services.FilesArtifactID]string
 }
 
@@ -41,7 +40,6 @@ func NewTestConfigurationBuilder() *TestConfigurationBuilder {
 		setupTimeoutSeconds: defaultSetupTimeoutSeconds,
 		runTimeoutSeconds:   defaultRunTimeoutSeconds,
 		isPartioningEnabled: defaultPartitioningEnabled,
-		staticFileFilepaths: map[services.StaticFileID]string{},
 		filesArtifactUrls:   map[services.FilesArtifactID]string{},
 	}
 }
@@ -61,12 +59,6 @@ func (builder *TestConfigurationBuilder) WithPartitioningEnabled(isPartitioningE
 	return builder
 }
 
-func (builder *TestConfigurationBuilder) WithStaticFileFilepaths(staticFileFilepaths map[services.StaticFileID]string) *TestConfigurationBuilder {
-	// TODO defensive copy
-	builder.staticFileFilepaths = staticFileFilepaths
-	return builder
-}
-
 func (builder *TestConfigurationBuilder) WithFilesArtifactUrls(filesArtifactUrls map[services.FilesArtifactID]string) *TestConfigurationBuilder {
 	// TODO defensive copy
 	builder.filesArtifactUrls = filesArtifactUrls
@@ -78,7 +70,6 @@ func (builder TestConfigurationBuilder) Build() *TestConfiguration {
 		SetupTimeoutSeconds:   builder.setupTimeoutSeconds,
 		RunTimeoutSeconds:     builder.runTimeoutSeconds,
 		IsPartitioningEnabled: builder.isPartioningEnabled,
-		StaticFileFilepaths:   builder.staticFileFilepaths,
 		FilesArtifactUrls:     builder.filesArtifactUrls,
 	}
 }