+++ typescript/src/lib/testsuite/test_configuration_builder.ts
@@ -1,4 +1,4 @@
-import { StaticFileID, FilesArtifactID } from "kurtosis-core-api-lib";
+import { FilesArtifactID } from "kurtosis-core-api-lib";
 import { TestConfiguration } from "./test_configuration";
 
 // vvvvvvvvv Update the docs if you change these vvvvvvvvvvv
@@ -12,14 +12,12 @@ export class TestConfigurationBuilder {
     private setupTimeoutSeconds: number;
     private runTimeoutSeconds: number;
     private isPartioningEnabled: boolean;
-    private staticFileFilepaths: Map<StaticFileID, string>;
     private filesArtifactUrls: Map<FilesArtifactID, string>;
     
     constructor () {
         this.setupTimeoutSeconds = DEFAULT_SETUP_TIMEOUT_SECONDS;
         this.runTimeoutSeconds = DEFAULT_RUN_TIMEOUT_SECONDS;
         this.isPartioningEnabled = DEFAULT_PARTITIONING_ENABLED;
-        this.staticFileFilepaths = new Map();
         this.filesArtifactUrls = new Map();
     }
     
@@ -38,12 +36,6 @@ export class TestConfigurationBuilder {
         return this;
     }
     
-    public withStaticFileFilepaths(staticFileFilepaths: Map<StaticFileID, string>): TestConfigurationBuilder {
-        // TODO defensive copy
-        this.staticFileFilepaths = staticFileFilepaths
-        return this;
-    }
-    
     public withFilesArtifactUrls(filesArtifactUrls: Map<FilesArtifactID, string>): TestConfigurationBuilder {
         // TODO defensive copy
         this.filesArtifactUrls = filesArtifactUrls;
@@ -54,8 +46,7 @@ export class TestConfigurationBuilder {
         return new TestConfiguration(
             this.setupTimeoutSeconds, 
             this.runTimeoutSeconds, 
-            this.isPartioningEnabled, 
-            this.staticFileFilepaths, 
+            this.isPartioningEnabled,
             this.filesArtifactUrls);
     }