+++ docs/changelog.md
@@ -1,4 +1,11 @@
 # TBD
+### Changes
+* Upgraded to [Kurt Client API 0.17.0](https://github.com/kurtosis-tech/kurtosis-client/blob/develop/docs/changelog.md#0170)
+* Renamed `RegisterFiles` to `RegisterFilesArtifacts` in testsuite service API because now it doesn't register static files, it only registers the file artifacts
+
+### Breaking Changes
+* Removed `StaticFileFilepaths` from `TestConfiguration` object
+  * Users should manually create and copy static files into the service container with the help of the new `SharedDirectory` and `SharedFileObject` objects
 
 # 0.7.1
 ### Fixes